Changed the global int's to unsigned 8 bit bc ARM

+++ include/asignment.hpp
@@ -23,9 +23,9 @@ ADIDigitalOut lowerClampy(1);
 bool usercontrol = false;
 bool autonGo = false;
 bool sideAuto = false;
-int liftdir = 0;
-int clampLiftState = 0;
-int clampLowerState = 0;
-int mPwr = 0;
-int turn = 0;
+uint8_t liftdir = 0;
+uint8_t clampLiftState = 0;
+uint8_t clampLowerState = 0;
+uint8_t mPwr = 0;
+uint8_t turn = 0;
